requirements.txt
openai>=1.30.0
chromadb>=0.5.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
pillow>=10.0.0
gradio>=4.30.0
python-dotenv>=1.0.0
requests>=2.31.0
nltk>=3.8.1
tqdm>=4.66.0
